Initial rework to make NFFS more generic
This patch starts reorganization of NFFS to make it more generic, i.e.
easier to port for other OS-es.

* introduce glue layer to abstract memory and flash operations
* remove fs specific stuff from "core" implementation (this will be part
  of OS ports)
* make rich public API - basically anything exposed in nffs_priv.h is
  a public API for now since OS ports will use this to implement fs
  stuff; long term this will need some cleanup.

Eventually, this will be a repository compatible with Mynewt which can
be used there directly with ports available for Zephyr and possibly
other OS-es. It should be then possible to build "core" code unmodified
on any OS with glue layer implemented.